Force TeamSeeder to skip creating existing permissions
TeamSeeder previously created permissions that already existed, if the migration had ran twice. This commit forces the seeder to verify if the permission it's attempting to create already exists, and if not, it creates it and assigns them at the end.

THis prevents an error from appearing if these permissions were created manually. In this case, the seeder will only assign the permissions instead of creating them.